server/5_portfolio_engine/requirements.txt
numpy==2.2.6
pandas==2.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
